Move `PoolParams` into their own module
