+++ eras/shelley/impl/cardano-ledger-shelley.cabal
@@ -53,6 +53,7 @@ library
     Cardano.Ledger.Shelley.Metadata
     Cardano.Ledger.Shelley.Orphans
     Cardano.Ledger.Shelley.PoolRank
+    Cardano.Ledger.Shelley.PoolParams
     Cardano.Ledger.Shelley.PParams
     Cardano.Ledger.Shelley.Rewards
     Cardano.Ledger.Shelley.RewardProvenance
